@@ -14,8 +14,8 @@
     "@vercel/analytics": "^1.1.1",
     "dompurify": "^2.4.1",
     "html-react-parser": "^3.0.6",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-router-dom": "^6.4.1",
     "react-scripts": "^5.0.1",
     "web-vitals": "^0.2.4"
